Feat: 공모 배정 확정 로직 구현 및 디테일 점검
- 대상·가드: SubscriptionStatus.PAID만 조회(정렬: paidAt ASC, id ASC), 총수요 집계 후 신청 0/수요 0 → allocated(0) 처리.
- 배정 알고리즘: 비례배정( floor ) → lotSize 배수 보정 → 잔여 라운드로빈(step=lotSize 또는 1)으로 공정 분배.
- 저장·멱등·상태전환: existsBy(subscriptionId, roundNo)로 멱등 가드 후 IpoAllocation 일괄 저장, 배정합계로 offering.allocated(assignedFinal) 전환.

+++ mkx-platform/src/main/java/com/beyond/MKX/domain/ipo/offering/dto/IpoOfferingReqDTO.java
@@ -23,7 +23,7 @@ public class IpoOfferingReqDTO {
     /* 청약 단위 */
     @NotNull
     @Positive
-    private Long lotSize;
+    private Integer lotSize;
 
     /* 청약 시작 시간 */
     @NotNull
